+++ src/signed_distance_field/transform.rs
@@ -12,14 +12,14 @@ pub struct Translation<T: SignedDistance>(pub Box<T>, pub Vec3);
 
 impl<T: SignedDistance> SignedDistance for Rotation<T> {
     #[inline]
-    fn distance_from(&self, point: Vec3) -> f64 {
+    fn distance_from(&self, point: Vec3) -> f32 {
         self.0.distance_from(point.apply_rotation(self.1.inverse()))
     }
 }
 
 impl<T: SignedDistance> SignedDistance for Translation<T> {
     #[inline]
-    fn distance_from(&self, point: Vec3) -> f64 {
+    fn distance_from(&self, point: Vec3) -> f32 {
         self.0.distance_from(point.sub(self.1))
     }
 }
